Progress on norm of polar bodies
We tried using linear programming in order to get the norm of a vector
related to a polar body. The PointsToHalfSpaces functions takes a set of
points and returns the linear system (matrix + coefficients) that's
supposed to be entered into the linear programming function which is
used in NormOfPolarBody.
The Norm function seems to work (we think...?) for the case of 2-dim
rectangle and rhombus. Applying this for a cube in 3-dim seems to
explode. We think it might be connected to the fact that the matrix
returned from PointsToHalfSpaces has linear dependent rows. We might be
able to fix this by removing the redundancy....