Discarded old algorithm - used fmincon instead
Total revision.
Matlab has a built it constrained optimization algorithms.
rewrote the code to use them istead
They get as input the function, the constraints and thier gradients.
Syntax is quite convoluted and might be made cleaner with some
syntactic-sugar of the language i am not familiar with.
Also changes A2n to a sparse matrix to improve runtime.