Some order
I moved around some files as a first step towards a more ordered
project.
Note that the previous original.m file is now CapacityScript.m, and the
new original.m is almost identical to the one we have at the master
branch. This should be renamed after we merge with the main branch.