Added a function to generate simplex param.
This function is still a WIP. So far it would even break for n not equal
to 4, *but* the structure of the code alludes to the general case.
Meaning, it traces what needs to happen in the main for loop for the
case where it builds the last two vectors. Now we pretty much just need
to replace the hard coded numbers for i,j.
The final step of the loop might need further work, it would seem I
might be getting a bad sign some place.

Next version will be a complete function ready to be tested.