Finished simplex parametrization
This update includes a finished GenerateSimplexParametrization function.
Checked it for n=2, it seems to work fine.
It takes a long time to finish for n>=3...
Maybe we could somehow replace the for loops with matrix operations.

Also included a simple test script which then substitutes the parameters
with those that fit the regular simplex in R^4, it works as it should.