Capacity of symplectic ellipsoids
This code uses the discretization we used for the general capacity
function, but instead of working with general convex body this altered
function works only with symplectic ellipsoids, given by their spectrum.

The idea is to test the general algorithm for the simplest case when the
h_K function is very tame and nice to work with.

The results so far are... unexpectedly bad.