Added Features - Lagrangian Products
a) changed m in capacity to m=60
b) quoted out some unused test code in G and dG
c) PolytopeCapacity - is a file to write test scripts
d) centroid.m - is a code I got from the internet to find the center of
mass of a polytope:  (this is not equal to the average of the vertices)

http://www.mathworks.com/matlabcentral/fileexchange/8514-centroid-of-a-convex-n-dimensional-polyhedron
e) CentralizePolytope translates it s.t. center of mass is at 0
f) ExtremePointsOnly - remove non extreme vertices from the polytope
g) LagrangianProductPolytope - constructs a lagrangian product of two
polytopes- this is easy in our coordinates system since we put all the
q's before the p's

Remarks:
Testing with product of triangle with sides Pi and a square of length 1
(should be symplectomorphic to the unit ball in R^4) gave correct
results, the value of Pi.

centroid is throwing warnings and without the {QJ} option I added in
it's code it throws error about cosphericality of cocircularity of some
of the points - this needs to be investigated.