More Capacity options and a script
* Added the functions for Minkowski Sum. They are actually faulty, but I
wanted to keep the files for backup, might be useful some day.
* Added options for Capacity function:
'minksum' - input 'on' to use the minkowski sum functions.
'epsilon' - the radius of the ball to be minkowski summed with P. (Yeah,
I know, crappy choice of name for this argument)
'startingtraj' - The value should be a vector with size [2*n*m 1].
* Added Kyler's script to generating values of capacities on a grid.