New script and modifications
* Added a new script that calculates c(K-K)/c(K) for K the std simplex
in dimension n, where n=2,...,50.
* Changed the script that searches for optimal simplex by adding
"reruns", i.e., each Capacity is calculated "reruns" number of times.
This is because the results of the previous run showed that we need
higher precision when we want to check against Viterbo's conjecture.
-> Also lowered the volume threshold in this script, and lowered the
size of the maximal entry to be 10 (instead of 100).
* Added the results of the run I did on the modified search for optimal
simplex script.