Trigonometric Escapades
Trying to pretty much redo the project in terms of trigonometric
function approximation for the closed curves, instead of linear
approximations.
There's probably an error in the gradient of F since it doesn't converge
as well as the automatically calculated gradient does.
Other than that, the code works much slower and gives really bad
results.
We should go over this at least one more time to make sure we did
everything correctly before dropping this.

We also dabbled a bit in mathematica when trying to figure out the
correct formulas after the transition to trigonometric ideas, so we
saved the notebook in case we need to do more trig.